@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 0.0.1+1
 
 environment:
-  sdk: ">=2.16.0-134.1.beta <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -39,9 +39,8 @@ dependencies:
       url: https://github.com/effektio/flyer-flutter_chat_ui.git
   flutter_chat_types: ^3.3.2
   google_fonts: ^2.3.1
-  flutter_svg: ^1.0.3
+  flutter_svg: ^0.22.0
   duration: ^3.0.11
-  image_picker: ^0.8.5
   file_picker: ^4.5.1
   open_file: ^3.2.1
   cached_network_image: ^3.2.1
@@ -69,17 +68,25 @@ dependencies:
   flutter_lorem: ^2.0.0
   get_time_ago: ^1.1.5
   implicitly_animated_reorderable_list: ^0.4.2
-  cached_memory_image: ^1.3.1
   flutter_mentions: ^2.0.1
-  flutter_html: ^3.0.0-alpha.5
+  flutter_matrix_html: ^1.0.0
   flutter_icons_null_safety: ^1.1.0
+  connectivity_plus: ^3.0.2
+  overlay_support: ^2.1.0
+  flutter_keyboard_visibility: ^5.4.0
+  timelines: ^0.1.0
+  window_size:
+    git:
+      url: https://github.com/google/flutter-desktop-embedding.git
+      path: plugins/window_size
+  beamer: ^1.5.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   integration_test:
     sdk: flutter
-  icons_launcher: ^2.0.5
+  icons_launcher: ^2.0.1
   flutter_lints: ^1.0.4
 
 
