@@ -1,18 +1,24 @@
 import 'dart:async';
+import 'dart:convert';
 import 'dart:io';
+import 'dart:typed_data';
 
+import 'package:beamer/beamer.dart';
 import 'package:effektio/controllers/receipt_controller.dart';
-import 'package:effektio/screens/HomeScreens/chat/ImageSelectionScreen.dart';
+import 'package:effektio/models/ImageSelectionModel.dart';
 import 'package:effektio/widgets/AppCommon.dart';
 import 'package:effektio_flutter_sdk/effektio_flutter_sdk_ffi.dart'
     show
         Client,
         Conversation,
         FfiBufferUint8,
-        FileDescription,
-        ImageDescription,
+        FileDesc,
+        ImageDesc,
         Member,
+        RoomEventItem,
         RoomMessage,
+        TextDesc,
+        TimelineDiff,
         TimelineStream,
         UserProfile;
 import 'package:file_picker/file_picker.dart';
@@ -21,7 +27,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
 import 'package:flutter_mentions/flutter_mentions.dart';
 import 'package:get/get.dart';
-import 'package:image_picker/image_picker.dart';
 import 'package:mime/mime.dart';
 import 'package:open_file/open_file.dart';
 import 'package:path_provider/path_provider.dart';
@@ -29,7 +34,8 @@ import 'package:permission_handler/permission_handler.dart';
 
 class ChatRoomController extends GetxController {
   Client client;
-  List<types.Message> messages = [];
+  late String userId;
+  final List<types.Message> _messages = [];
   List<types.User> typingUsers = [];
   TimelineStream? _stream;
   RxBool isLoading = false.obs;
@@ -42,21 +48,29 @@ class ChatRoomController extends GetxController {
   GlobalKey<FlutterMentionsState> mentionKey =
       GlobalKey<FlutterMentionsState>();
   bool isSendButtonVisible = false;
-  final List<XFile> _imageFileList = [];
+  bool isEmojiContainerVisible = false;
+  final List<PlatformFile> _imageFileList = [];
   List<Member> activeMembers = [];
   Map<String, String> messageTextMapMarkDown = {};
   Map<String, String> messageTextMapHtml = {};
   final Map<String, Future<FfiBufferUint8>> _userAvatars = {};
   final Map<String, String> _userNames = {};
   List<Map<String, dynamic>> mentionList = [];
+  StreamSubscription<TimelineDiff>? _diffSubscription;
   StreamSubscription<RoomMessage>? _messageSubscription;
+  int emojiMessageIndex = 0;
+  String? emojiCurrentId;
+  String? authorId;
+  bool showReplyView = false;
+  Widget? replyMessageWidget;
+  types.Message? repliedToMessage;
 
   ChatRoomController({required this.client}) : super();
 
   @override
   void onInit() {
     super.onInit();
-
+    userId = client.userId().toString();
     focusNode.addListener(() {
       if (focusNode.hasFocus) {
         isEmojiVisible.value = false;
@@ -64,7 +78,7 @@ class ChatRoomController extends GetxController {
       }
     });
 
-    _messageSubscription = client.incomingMessageRx()?.listen((event) {
+    _messageSubscription = client.incomingMessageRx()?.listen((event) async {
       // the latest message is dealt in convo receiver of ChatListController
       // here manage only its message history
       if (_currentRoom != null) {
@@ -72,9 +86,26 @@ class ChatRoomController extends GetxController {
         if (event.roomId() == _currentRoom!.getRoomId()) {
           // filter only message from other not me
           // it is processed in handleSendPressed
-          if (event.sender() != client.userId().toString()) {
-            _loadMessage(event);
-            update(['Chat']);
+          types.Message m = await _prepareMessage(event);
+          if (m is! types.UnsupportedMessage) {
+            _insertMessage(m);
+            RoomEventItem? eventItem = event.eventItem();
+            if (eventItem != null) {
+              if (eventItem.sender() != client.userId().toString()) {
+                if (isLoading.isFalse) {
+                  update(['Chat']);
+                }
+                if (eventItem.msgtype() == 'm.image') {
+                  _fetchMessageContent(m.id);
+                }
+              }
+              if (m.metadata != null && m.metadata!.containsKey('repliedTo')) {
+                _fetchOriginalContent(
+                  m.metadata?['repliedTo'],
+                  m.id,
+                );
+              }
+            }
           }
         }
       }
@@ -84,6 +115,7 @@ class ChatRoomController extends GetxController {
   @override
   void onClose() {
     focusNode.removeListener(() {});
+    _diffSubscription?.cancel();
     _messageSubscription?.cancel();
 
     super.onClose();
@@ -92,10 +124,11 @@ class ChatRoomController extends GetxController {
   //get the timeline of room
   Future<void> setCurrentRoom(Conversation? convoRoom) async {
     if (convoRoom == null) {
-      messages.clear();
+      _messages.clear();
       typingUsers.clear();
       activeMembers.clear();
       mentionList.clear();
+      _diffSubscription?.cancel();
       _stream = null;
       _page = 0;
       _currentRoom = null;
@@ -103,7 +136,7 @@ class ChatRoomController extends GetxController {
       _currentRoom = convoRoom;
       update(['room-profile']);
       isLoading.value = true;
-      activeMembers = (await _currentRoom!.activeMembers()).toList();
+      activeMembers = (await convoRoom.activeMembers()).toList();
       update(['active-members']);
       _fetchUserProfiles();
       if (_currentRoom == null) {
@@ -111,17 +144,171 @@ class ChatRoomController extends GetxController {
         isLoading.value = false;
         return;
       }
-      _stream = _currentRoom!.timeline();
-      // i am fetching messages from remote
+      _stream = await _currentRoom!.timelineStream();
+      // event handler from paginate
+      _diffSubscription = _stream?.diffRx().listen((event) async {
+        switch (event.action()) {
+          case 'Replace':
+            debugPrint('chat room message replace');
+            List<RoomMessage> values = event.values()!.toList();
+            for (RoomMessage msg in values) {
+              types.Message m = await _prepareMessage(msg);
+              if (m is! types.UnsupportedMessage) {
+                _insertMessage(m);
+                if (m.metadata != null &&
+                    m.metadata!.containsKey('repliedTo')) {
+                  _fetchOriginalContent(
+                    m.metadata?['repliedTo']['eventId'],
+                    m.id,
+                  );
+                }
+                RoomEventItem? eventItem = msg.eventItem();
+                if (eventItem != null) {
+                  if (isLoading.isFalse) {
+                    update(['Chat']);
+                  }
+                  if (eventItem.msgtype() == 'm.image') {
+                    _fetchMessageContent(m.id);
+                  }
+                }
+              }
+            }
+            break;
+          case 'InsertAt':
+            debugPrint('chat room message insert at');
+            RoomMessage value = event.value()!;
+            types.Message m = await _prepareMessage(value);
+            if (m is! types.UnsupportedMessage) {
+              _insertMessage(m);
+              if (m.metadata != null && m.metadata!.containsKey('repliedTo')) {
+                _fetchOriginalContent(
+                  m.metadata?['repliedTo'],
+                  m.id,
+                );
+              }
+              RoomEventItem? eventItem = value.eventItem();
+              if (eventItem != null) {
+                if (isLoading.isFalse) {
+                  update(['Chat']);
+                }
+                if (eventItem.msgtype() == 'm.image') {
+                  _fetchMessageContent(m.id);
+                }
+              }
+            }
+            break;
+          case 'UpdateAt':
+            debugPrint('chat room message update at');
+            RoomMessage value = event.value()!;
+            types.Message m = await _prepareMessage(value);
+            if (m is! types.UnsupportedMessage) {
+              _updateMessage(m);
+              if (m.metadata != null && m.metadata!.containsKey('repliedTo')) {
+                _fetchOriginalContent(
+                  m.metadata?['repliedTo'],
+                  m.id,
+                );
+              }
+              RoomEventItem? eventItem = value.eventItem();
+              if (eventItem != null) {
+                if (isLoading.isFalse) {
+                  update(['Chat']);
+                }
+                RoomEventItem? eventItem = value.eventItem();
+                if (eventItem != null) {
+                  if (isLoading.isFalse) {
+                    update(['Chat']);
+                  }
+                  if (eventItem.msgtype() == 'm.image') {
+                    _fetchMessageContent(m.id);
+                  }
+                }
+              }
+            }
+            break;
+          case 'Push':
+            debugPrint('chat room message push');
+            RoomMessage value = event.value()!;
+            types.Message m = await _prepareMessage(value);
+            if (m is! types.UnsupportedMessage) {
+              _messages.insert(0, m);
+              if (m.metadata != null && m.metadata!.containsKey('repliedTo')) {
+                _fetchOriginalContent(
+                  m.metadata?['repliedTo'],
+                  m.id,
+                );
+              }
+              RoomEventItem? eventItem = value.eventItem();
+              if (eventItem != null) {
+                if (isLoading.isFalse) {
+                  update(['Chat']);
+                }
+                RoomEventItem? eventItem = value.eventItem();
+                if (eventItem != null) {
+                  if (isLoading.isFalse) {
+                    update(['Chat']);
+                  }
+                  if (eventItem.msgtype() == 'm.image') {
+                    _fetchMessageContent(m.id);
+                  }
+                }
+              }
+            }
+            break;
+          case 'RemoveAt':
+            int index = event.index()!;
+            _messages.removeAt(_messages.length - index);
+            if (isLoading.isFalse) {
+              update(['Chat']);
+            }
+            break;
+          case 'Move':
+            int oldIndex = event.oldIndex()!;
+            int newIndex = event.newIndex()!;
+            int i = _messages.length - newIndex;
+            if (oldIndex < newIndex) {
+              i += 1;
+            }
+            types.Message m = _messages.removeAt(_messages.length - oldIndex);
+            if (m is! types.UnsupportedMessage) {
+              _messages.insert(i, m);
+              if (m.metadata != null && m.metadata!.containsKey('repliedTo')) {
+                _fetchOriginalContent(
+                  m.metadata?['repliedTo'],
+                  m.id,
+                );
+              }
+              if (isLoading.isFalse) {
+                update(['Chat']);
+              }
+            }
+            break;
+          case 'Pop':
+            _messages.removeLast();
+            if (isLoading.isFalse) {
+              update(['Chat']);
+            }
+            break;
+          case 'Clear':
+            _messages.clear();
+            if (isLoading.isFalse) {
+              update(['Chat']);
+            }
+            break;
+        }
+      });
+
       if (_currentRoom == null) {
         // user may close chat screen before long loading completed
         isLoading.value = false;
         return;
       }
-      var msgs = await _stream!.paginateBackwards(10);
-      for (RoomMessage message in msgs) {
-        _loadMessage(message);
-      }
+      bool hasMore = true;
+      do {
+        hasMore = await _stream!.paginateBackwards(10);
+        // wait for diff rx to be finished
+        sleep(const Duration(milliseconds: 500));
+      } while (hasMore && _messages.length < 10);
       // load receipt status of room
       var receiptController = Get.find<ReceiptController>();
       var receipts = (await convoRoom.userReceipts()).toList();
@@ -187,12 +374,13 @@ class ChatRoomController extends GetxController {
     types.TextMessage message,
     types.PreviewData previewData,
   ) {
-    final idx = messages.indexWhere((element) => element.id == message.id);
-    final updatedMessage =
-        (messages[idx] as types.TextMessage).copyWith(previewData: previewData);
+    final index = _messages.indexWhere((x) => x.id == message.id);
+    final updatedMessage = (_messages[index] as types.TextMessage).copyWith(
+      previewData: previewData,
+    );
 
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      messages[idx] = updatedMessage;
+    WidgetsBinding.instance.addPostFrameCallback((Duration duration) {
+      _messages[index] = updatedMessage;
       update(['Chat']);
     });
   }
@@ -206,20 +394,19 @@ class ChatRoomController extends GetxController {
     // image or video is sent automatically
     // user will click "send" button explicitly for text only
     await _currentRoom!.typingNotice(false);
-    var eventId = await _currentRoom!.sendFormattedMessage(markdownMessage);
-    final textMessage = types.TextMessage(
-      author: types.User(id: client.userId().toString()),
-      createdAt: DateTime.now().millisecondsSinceEpoch,
-      id: eventId,
-      text: htmlMessage,
-      status: types.Status.sent,
-      showStatus: true,
-      metadata: {
-        'messageLength': messageLength,
-      },
-    );
-    messages.insert(0, textMessage);
-    update(['Chat']);
+    if (repliedToMessage != null) {
+      await _currentRoom!.sendTextReply(
+        markdownMessage,
+        repliedToMessage!.id,
+        null,
+      );
+      repliedToMessage = null;
+      replyMessageWidget = null;
+      showReplyView = false;
+      update(['chat-input']);
+    } else {
+      await _currentRoom!.sendFormattedMessage(markdownMessage);
+    }
   }
 
   Future<void> handleMultipleImageSelection(
@@ -227,118 +414,133 @@ class ChatRoomController extends GetxController {
     String roomName,
   ) async {
     _imageFileList.clear();
-    final result = await ImagePicker().pickMultiImage(
-      imageQuality: 70,
-      maxWidth: 1440,
+    FilePickerResult? result = await FilePicker.platform.pickFiles(
+      type: FileType.image,
+      allowMultiple: true,
     );
-    if (result != null) {
-      _imageFileList.addAll(result);
-      Navigator.push(
-        context,
-        MaterialPageRoute(
-          builder: (context) => ImageSelection(
-            imageList: _imageFileList,
-            roomName: roomName,
-          ),
-        ),
-      );
+    if (result == null) {
+      return;
     }
+    _imageFileList.addAll(result.files);
+    Beamer.of(context).beamToNamed('/imageSelection', data: ImageSelectionModel(imageList: _imageFileList, roomName: roomName));
   }
 
-  Future<void> sendImage(XFile? result) async {
-    if (result != null) {
-      final bytes = await result.readAsBytes();
-      final image = await decodeImageFromList(bytes);
-      final mimeType = lookupMimeType(result.path);
-      var eventId = await _currentRoom!.sendImageMessage(
-        result.path,
-        result.name,
+  Future<void> sendImage(PlatformFile file) async {
+    String? path = file.path;
+    if (path == null) {
+      return;
+    }
+    String? name = file.name;
+    String? mimeType = lookupMimeType(path);
+    Uint8List? bytes = file.bytes;
+    if (bytes == null) {
+      return;
+    }
+    final image = await decodeImageFromList(bytes);
+    if (repliedToMessage != null) {
+      await _currentRoom!.sendImageReply(
+        path,
+        name,
+        mimeType!,
+        bytes.length,
+        image.width,
+        image.height,
+        repliedToMessage!.id,
+        null,
+      );
+      repliedToMessage = null;
+      replyMessageWidget = null;
+      showReplyView = false;
+      update(['chat-input']);
+    } else {
+      await _currentRoom!.sendImageMessage(
+        path,
+        name,
         mimeType!,
         bytes.length,
         image.width,
         image.height,
       );
-
-      final message = types.ImageMessage(
-        author: types.User(id: client.userId().toString()),
-        createdAt: DateTime.now().millisecondsSinceEpoch,
-        height: image.height.toDouble(),
-        id: eventId,
-        name: result.name,
-        size: bytes.length,
-        uri: result.path,
-        width: image.width.toDouble(),
-      );
-      messages.insert(0, message);
-      update(['Chat']);
     }
   }
 
   //image selection
   Future<void> handleImageSelection(BuildContext context) async {
-    final result = await ImagePicker().pickImage(
-      imageQuality: 70,
-      maxWidth: 1440,
-      source: ImageSource.gallery,
+    FilePickerResult? result = await FilePicker.platform.pickFiles(
+      type: FileType.image,
     );
-
-    if (result != null) {
-      final bytes = await result.readAsBytes();
-      final image = await decodeImageFromList(bytes);
-      final mimeType = lookupMimeType(result.path);
-      var eventId = await _currentRoom!.sendImageMessage(
-        result.path,
-        result.name,
+    if (result == null) {
+      return;
+    }
+    String? path = result.files.single.path;
+    if (path == null) {
+      return;
+    }
+    String? name = result.files.single.name;
+    String? mimeType = lookupMimeType(path);
+    Uint8List bytes = File(path).readAsBytesSync();
+    final image = await decodeImageFromList(bytes);
+    if (repliedToMessage != null) {
+      await _currentRoom!.sendImageReply(
+        path,
+        name,
+        mimeType!,
+        bytes.length,
+        image.width,
+        image.height,
+        repliedToMessage!.id,
+        null,
+      );
+      repliedToMessage = null;
+      replyMessageWidget = null;
+      showReplyView = false;
+      update(['chat-input']);
+    } else {
+      await _currentRoom!.sendImageMessage(
+        path,
+        name,
         mimeType!,
         bytes.length,
         image.width,
         image.height,
       );
-
-      // i am sending message
-      final message = types.ImageMessage(
-        author: types.User(id: client.userId().toString()),
-        createdAt: DateTime.now().millisecondsSinceEpoch,
-        height: image.height.toDouble(),
-        id: eventId,
-        name: result.name,
-        size: bytes.length,
-        uri: result.path,
-        width: image.width.toDouble(),
-      );
-      Navigator.pop(context);
-      messages.insert(0, message);
-      update(['Chat']);
     }
   }
 
   //file selection
   Future<void> handleFileSelection(BuildContext context) async {
-    final result = await FilePicker.platform.pickFiles(
+    FilePickerResult? result = await FilePicker.platform.pickFiles(
       type: FileType.any,
     );
-
-    if (result != null && result.files.single.path != null) {
-      final mimeType = lookupMimeType(result.files.single.path!);
+    if (result == null) {
+      return;
+    }
+    String? path = result.files.single.path;
+    if (path == null) {
+      return;
+    }
+    String? name = result.files.single.name;
+    String? mimeType = lookupMimeType(path);
+    if (repliedToMessage != null) {
+      await _currentRoom!.sendFileReply(
+        path,
+        name,
+        mimeType!,
+        result.files.single.size,
+        repliedToMessage!.id,
+        null,
+      );
+      repliedToMessage = null;
+      replyMessageWidget = null;
+      showReplyView = false;
+      update(['chat-input']);
+    } else {
       await _currentRoom!.sendFileMessage(
-        result.files.single.path!,
-        result.files.single.name,
+        path,
+        name,
         mimeType!,
         result.files.single.size,
       );
-
-      // i am sending message
-      final message = types.FileMessage(
-        author: types.User(id: client.userId().toString()),
-        createdAt: DateTime.now().millisecondsSinceEpoch,
-        id: randomString(),
-        name: result.files.single.name,
-        size: result.files.single.size,
-        uri: result.files.single.path!,
-      );
-      Navigator.pop(context);
-      messages.insert(0, message);
-      update(['Chat']);
     }
   }
 
@@ -379,116 +581,268 @@ class ChatRoomController extends GetxController {
 
   //Pagination Control
   Future<void> handleEndReached() async {
-    final msgs = await _stream!.paginateBackwards(10);
-    // i am fetching messages from remote
-    for (RoomMessage message in msgs) {
-      _loadMessage(message);
-    }
+    bool hasMore = await _stream!.paginateBackwards(10);
+    debugPrint('backward pagination has more: $hasMore');
     _page = _page + 1;
     update(['Chat']);
   }
 
   void _insertMessage(types.Message m) {
     var receiptController = Get.find<ReceiptController>();
+    if (m.type != types.MessageType.unsupported &&
+        m.type != types.MessageType.custom) {
+      List<String> seenByList = receiptController.getSeenByList(
+        _currentRoom!.getRoomId(),
+        m.createdAt!,
+      );
+      if (m.author.id == client.userId().toString()) {
+        types.Status status = seenByList.isEmpty
+            ? types.Status.sent
+            : seenByList.length < activeMembers.length
+                ? types.Status.delivered
+                : types.Status.seen;
+        _messages.add(m.copyWith(showStatus: true, status: status));
+        return;
+      }
+    }
+    _messages.add(m);
+  }
+
+  void _updateMessage(types.Message m) {
+    var receiptController = Get.find<ReceiptController>();
+    int index = _messages.indexWhere((msg) => m.id == msg.id);
     List<String> seenByList = receiptController.getSeenByList(
       _currentRoom!.getRoomId(),
       m.createdAt!,
     );
-    var msg = (m.author.id == client.userId().toString())
-        ? m.copyWith(
-            showStatus: true,
-            status: seenByList.length < activeMembers.length
-                ? types.Status.delivered
-                : types.Status.seen,
-          )
-        : m;
-    for (var i = 0; i < messages.length; i++) {
-      if (messages[i].createdAt! < m.createdAt!) {
-        messages.insert(i, msg);
-        return;
-      }
+    if (m.author.id == client.userId().toString()) {
+      types.Status status = seenByList.isEmpty
+          ? types.Status.sent
+          : seenByList.length < activeMembers.length
+              ? types.Status.delivered
+              : types.Status.seen;
+      _messages[index] = m.copyWith(showStatus: true, status: status);
+      return;
     }
-    messages.add(msg);
+    _messages[index] = m;
   }
 
-  void _loadMessage(RoomMessage message) {
-    String msgtype = message.msgtype();
-    String sender = message.sender();
+  Future<types.Message> _prepareMessage(RoomMessage message) async {
+    RoomEventItem? eventItem = message.eventItem();
+    if (eventItem == null) {
+      // should not return null, before we can keep track of index in diff receiver
+      return types.UnsupportedMessage(
+        createdAt: DateTime.now().millisecondsSinceEpoch,
+        author: types.User(id: client.userId().toString()),
+        id: UniqueKey().toString(),
+        metadata: const {
+          'itemType': 'virtual',
+        },
+      );
+    }
+
+    String? msgtype = eventItem.msgtype();
+    String sender = eventItem.sender();
     var author = types.User(id: sender, firstName: simplifyUserId(sender));
-    int? createdAt = message.originServerTs(); // in milliseconds
-    String eventId = message.eventId();
+    int? createdAt = eventItem.originServerTs(); // in milliseconds
+    String eventId = eventItem.eventId();
+
+    String? inReplyTo = eventItem.inReplyTo();
+
+    Map<String, dynamic> reactions = {};
+    for (var key in eventItem.reactionKeys()) {
+      String k = key.toDartString();
+      reactions[k] = eventItem.reactionDesc(k);
+    }
 
     if (msgtype == 'm.audio') {
     } else if (msgtype == 'm.emote') {
     } else if (msgtype == 'm.file') {
-      FileDescription? description = message.fileDescription();
+      FileDesc? description = eventItem.fileDesc();
       if (description != null) {
-        types.FileMessage m = types.FileMessage(
+        Map<String, dynamic> metadata = {};
+        if (inReplyTo != null) {
+          metadata['repliedTo'] = inReplyTo;
+        }
+        if (reactions.isNotEmpty) {
+          metadata['reactions'] = reactions;
+        }
+        return types.FileMessage(
           author: author,
           createdAt: createdAt,
           id: eventId,
+          metadata: metadata,
           name: description.name(),
           size: description.size() ?? 0,
           uri: '',
         );
-        _insertMessage(m);
-      }
-      if (isLoading.isFalse) {
-        update(['Chat']);
       }
     } else if (msgtype == 'm.image') {
-      ImageDescription? description = message.imageDescription();
+      ImageDesc? description = eventItem.imageDesc();
       if (description != null) {
-        types.ImageMessage m = types.ImageMessage(
+        Map<String, dynamic> metadata = {};
+        if (inReplyTo != null) {
+          metadata['repliedTo'] = inReplyTo;
+        }
+        if (reactions.isNotEmpty) {
+          metadata['reactions'] = reactions;
+        }
+        metadata['base64'] = '';
+        return types.ImageMessage(
           author: author,
           createdAt: createdAt,
           height: description.height()?.toDouble(),
           id: eventId,
+          metadata: metadata,
           name: description.name(),
           size: description.size() ?? 0,
           uri: '',
           width: description.width()?.toDouble(),
         );
-        _insertMessage(m);
-        if (isLoading.isFalse) {
-          update(['Chat']);
-        }
-        _currentRoom!.imageBinary(eventId).then((data) {
-          int idx = messages.indexWhere((x) => x.id == eventId);
-          if (idx != -1) {
-            messages[idx] = messages[idx].copyWith(
-              metadata: {
-                'binary': data.asTypedList(),
-              },
-            );
-            update(['Chat']);
-          }
-        });
       }
     } else if (msgtype == 'm.location') {
     } else if (msgtype == 'm.notice') {
     } else if (msgtype == 'm.server_notice') {
     } else if (msgtype == 'm.text') {
-      types.TextMessage m = types.TextMessage(
-        author: author,
-        createdAt: createdAt,
-        id: eventId,
-        text: message.formattedBody() ?? message.body(),
-        metadata: {
-          'messageLength': message.body().length,
-        },
-      );
-      _insertMessage(m);
+      TextDesc? description = eventItem.textDesc();
+      if (description != null) {
+        String? formattedBody = description.formattedBody();
+        String body = description.body(); // always exists
+        Map<String, dynamic> metadata = {
+          'messageLength': body.length,
+        };
+        if (inReplyTo != null) {
+          metadata['repliedTo'] = inReplyTo;
+        }
+        if (reactions.isNotEmpty) {
+          metadata['reactions'] = reactions;
+        }
+        //check whether string only contains emoji(s).
+        if (isOnlyEmojis(description.body())) {
+          metadata['enlargeEmoji'] = true;
+        } else {
+          metadata['enlargeEmoji'] = false;
+        }
+
+        return types.TextMessage(
+          author: author,
+          createdAt: createdAt,
+          id: eventId,
+          metadata: metadata,
+          text: formattedBody ?? body,
+        );
+      }
+    } else if (msgtype == 'm.video') {
+    } else if (msgtype == 'm.key.verification.request') {
+    } else if (msgtype == 'm.sticker') {}
+
+    // should not return null, before we can keep track of index in diff receiver
+    return types.CustomMessage(
+      author: author,
+      createdAt: createdAt,
+      id: eventId,
+      metadata: {
+        'itemType': 'event',
+        'itemContentType': eventItem.itemContentType(),
+      },
+    );
+  }
+
+  List<types.Message> getMessages() {
+    List<types.Message> msgs = _messages.where((x) {
+      if (x.metadata?['itemType'] == 'virtual') {
+        // UnsupportedMessage
+        return false;
+      }
+      return true;
+    }).toList();
+    return msgs;
+  }
+
+  void _fetchMessageContent(String eventId) {
+    _currentRoom!.imageBinary(eventId).then((data) {
+      int index = _messages.indexWhere((x) => x.id == eventId);
+      if (index != -1) {
+        final metadata = _messages[index].metadata ?? {};
+        metadata['base64'] = base64Encode(data.asTypedList());
+        _messages[index] = _messages[index].copyWith(metadata: metadata);
+        if (isLoading.isFalse) {
+          update(['Chat']);
+        }
+      }
+    });
+  }
+
+  // fetch original content media for reply msg .i.e. text,image,file etc.
+  void _fetchOriginalContent(String originalId, String replyId) {
+    _currentRoom!.getMessage(originalId).then((roomMsg) {
+      // reply is allowed for only EventItem not VirtualItem
+      // user should be able to get original event as RoomMessage
+      RoomEventItem orgEventItem = roomMsg.eventItem()!;
+      String? orgMsgType = orgEventItem.msgtype();
+      Map<String, dynamic> repliedToContent = {};
+      types.Message? repliedTo;
+      if (orgMsgType == 'm.text') {
+        repliedToContent = {
+          'content': orgEventItem.textDesc()!.body(),
+          'messageLength': orgEventItem.textDesc()!.body().length,
+        };
+        repliedTo = types.TextMessage(
+          author: types.User(id: orgEventItem.sender()),
+          id: originalId,
+          createdAt: orgEventItem.originServerTs(),
+          text: orgEventItem.textDesc()!.body(),
+          metadata: repliedToContent,
+        );
+      } else if (orgMsgType == 'm.image') {
+        _currentRoom!.imageBinary(originalId).then((data) {
+          repliedToContent['content'] = base64Encode(data.asTypedList());
+        });
+        repliedTo = types.ImageMessage(
+          author: types.User(id: orgEventItem.sender()),
+          id: originalId,
+          createdAt: orgEventItem.originServerTs(),
+          name: orgEventItem.imageDesc()!.name(),
+          size: orgEventItem.imageDesc()!.size()!,
+          uri: '',
+          metadata: repliedToContent,
+        );
+      } else if (orgMsgType == 'm.file') {
+        repliedToContent = {
+          'content': orgEventItem.fileDesc()!.name(),
+        };
+        repliedTo = types.FileMessage(
+          author: types.User(id: orgEventItem.sender()),
+          id: originalId,
+          createdAt: orgEventItem.originServerTs(),
+          name: orgEventItem.fileDesc()!.name(),
+          size: orgEventItem.fileDesc()!.size()!,
+          uri: '',
+          metadata: repliedToContent,
+        );
+      } else if (orgMsgType == 'm.sticker') {
+        // user can't do any action about sticker message
+      }
+      int index = _messages.indexWhere((x) => x.id == replyId);
+      if (index != -1) {
+        _messages[index] = _messages[index].copyWith(repliedMessage: repliedTo);
+      }
       if (isLoading.isFalse) {
         update(['Chat']);
       }
-    } else if (msgtype == 'm.video') {
-    } else if (msgtype == 'm.key.verification.request') {}
+    });
   }
 
+  /// Update button state based on text editor.
   void sendButtonUpdate() {
     isSendButtonVisible =
         mentionKey.currentState!.controller!.text.trim().isNotEmpty;
+    update(['chat-input']);
+  }
+
+  /// Disable button as soon as send button is pressed.
+  void sendButtonDisable() {
+    isSendButtonVisible = !isSendButtonVisible;
     update(['chat-input']);
   }
 
@@ -497,5 +851,35 @@ class ChatRoomController extends GetxController {
       return Future.value(false);
     }
     return await _currentRoom!.typingNotice(typing);
+  }
+
+  void updateEmojiState(types.Message message) {
+    emojiMessageIndex = _messages.indexWhere((x) => x.id == message.id);
+    emojiCurrentId = _messages[emojiMessageIndex].id;
+    if (emojiCurrentId == message.id) {
+      isEmojiContainerVisible = !isEmojiContainerVisible;
+    }
+
+    if (isEmojiContainerVisible) {
+      authorId = message.author.id;
+    }
+    update(['emoji-reaction']);
+  }
+
+  bool isAuthor() {
+    return client.userId().toString() == authorId;
+  }
+
+  void toggleEmojiContainer() {
+    isEmojiContainerVisible = !isEmojiContainerVisible;
+    update(['emoji-reaction']);
+  }
+
+  Future<void> sendEmojiReaction(String eventId, String emoji) async {
+    await _currentRoom!.sendReaction(eventId, emoji);
+  }
+
+  Future<void> redactRoomMessage(String eventId) async {
+    await _currentRoom!.redactMessage(eventId, '', null);
   }
 }
