@@ -1,8 +1,11 @@
 import 'dart:io';
 
+import 'package:beamer/beamer.dart';
 import 'package:effektio/common/store/themes/SeperatedThemes.dart';
 import 'package:effektio/controllers/chat_room_controller.dart';
+import 'package:effektio/models/ImageSelectionModel.dart';
 import 'package:effektio/widgets/CustomChatInput.dart';
+import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:get/get.dart';
@@ -10,13 +13,11 @@ import 'package:photo_view/photo_view.dart';
 import 'package:photo_view/photo_view_gallery.dart';
 
 class ImageSelection extends StatefulWidget {
-  final List imageList;
-  final String roomName;
+ final ImageSelectionModel imageSelectionModel;
 
   const ImageSelection({
     Key? key,
-    required this.imageList,
-    required this.roomName,
+    required this.imageSelectionModel,
   }) : super(key: key);
 
   @override
@@ -36,11 +37,12 @@ class _ImageSelectionState extends State<ImageSelection> {
         children: [
           Expanded(
             child: PhotoViewGallery.builder(
-              itemCount: widget.imageList.length,
+              itemCount: widget.imageSelectionModel.imageList.length,
               scrollPhysics: const BouncingScrollPhysics(),
               builder: (context, index) {
+                PlatformFile file = widget.imageSelectionModel.imageList[index];
                 return PhotoViewGalleryPageOptions(
-                  imageProvider: FileImage(File(widget.imageList[index].path)),
+                  imageProvider: FileImage(File(file.path!)),
                   initialScale: PhotoViewComputedScale.contained * 0.8,
                 );
               },
@@ -60,10 +62,10 @@ class _ImageSelectionState extends State<ImageSelection> {
               scrollDirection: Axis.horizontal,
               child: Row(
                 children: [
-                  for (var item in widget.imageList)
+                  for (PlatformFile file in widget.imageSelectionModel.imageList)
                     Padding(
                       padding: const EdgeInsets.only(top: 10, left: 8),
-                      child: buildImageItem(item),
+                      child: buildImageItem(file),
                     ),
                   Padding(
                     padding: const EdgeInsets.symmetric(horizontal: 30),
@@ -77,11 +79,11 @@ class _ImageSelectionState extends State<ImageSelection> {
           ),
           CustomChatInput(
             isChatScreen: false,
-            roomName: widget.roomName,
+            roomName: widget.imageSelectionModel.roomName,
             onButtonPressed: () async {
-              Navigator.of(context).pop();
-              for (var image in widget.imageList) {
-                await controller.sendImage(image);
+              Beamer.of(context).beamBack();
+              for (PlatformFile file in widget.imageSelectionModel.imageList) {
+                await controller.sendImage(file);
               }
             },
           )
@@ -90,11 +92,11 @@ class _ImageSelectionState extends State<ImageSelection> {
     );
   }
 
-  Widget buildImageItem(item) {
+  Widget buildImageItem(PlatformFile file) {
     return InkWell(
       onTap: () {
         setState(() {
-          selectedIndex = widget.imageList.indexOf(item);
+          selectedIndex = widget.imageSelectionModel.imageList.indexOf(file);
           pageController.jumpToPage(selectedIndex);
         });
       },
@@ -107,10 +109,10 @@ class _ImageSelectionState extends State<ImageSelection> {
               borderRadius: BorderRadius.circular(3),
               image: DecorationImage(
                 fit: BoxFit.fill,
-                image: FileImage(File(item.path)),
+                image: FileImage(File(file.path!)),
               ),
               color: AppCommonTheme.backgroundColor,
-              border: getItemBorder(item),
+              border: getItemBorder(file),
             ),
           ),
           Positioned(
@@ -118,10 +120,10 @@ class _ImageSelectionState extends State<ImageSelection> {
             top: 4,
             child: InkWell(
               onTap: () {
-                int idx = widget.imageList.indexOf(item);
-                widget.imageList.removeAt(idx);
-                if (widget.imageList.isEmpty) {
-                  Navigator.of(context).pop();
+                int idx = widget.imageSelectionModel.imageList.indexOf(file);
+                widget.imageSelectionModel.imageList.removeAt(idx);
+                if (widget.imageSelectionModel.imageList.isEmpty) {
+                  Beamer.of(context).beamBack();
                 }
               },
               child: CircleAvatar(
@@ -138,8 +140,8 @@ class _ImageSelectionState extends State<ImageSelection> {
     );
   }
 
-  BoxBorder? getItemBorder(item) {
-    if (selectedIndex != widget.imageList.indexOf(item)) {
+  BoxBorder? getItemBorder(PlatformFile file) {
+    if (selectedIndex != widget.imageSelectionModel.imageList.indexOf(file)) {
       return null;
     }
     return Border.all(
