@@ -14,7 +14,6 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:get/get.dart';
 import 'package:implicitly_animated_reorderable_list/implicitly_animated_reorderable_list.dart';
 import 'package:implicitly_animated_reorderable_list/transitions.dart';
-// import 'package:themed/themed.dart';
 
 class ChatOverview extends StatefulWidget {
   final Client client;
@@ -26,6 +25,7 @@ class ChatOverview extends StatefulWidget {
 }
 
 class _ChatOverviewState extends State<ChatOverview> {
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -93,7 +93,10 @@ class _ChatOverviewState extends State<ChatOverview> {
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.center,
               children: [
-                if (widget.client.isGuest()) empty else buildList(context),
+                if (widget.client.isGuest())
+                  empty
+                else
+                  buildList(context),
               ],
             ),
           ),
@@ -234,20 +237,21 @@ class _ChatOverviewState extends State<ChatOverview> {
 
   Widget buildInvitedItem(Invitation item) {
     return InviteInfoWidget(
-      userId: widget.client.userId().toString(),
+      client: widget.client,
       invitation: item,
       avatarColor: Colors.white,
     );
   }
 
   Widget buildJoinedItem(JoinedRoom item) {
+    String roomId = item.conversation.getRoomId();
     // we should be able to update only changed room items
     // so we use GetBuilder to render item
     return GetBuilder<ChatListController>(
-      id: 'chatroom-${item.conversation.getRoomId()}',
+      id: 'chatroom-$roomId',
       builder: (controller) => ChatListItem(
-        key: Key(item.conversation.getRoomId()),
-        userId: widget.client.userId().toString(),
+        key: Key(roomId),
+        client: widget.client,
         room: item.conversation,
         latestMessage: item.latestMessage,
         typingUsers: item.typingUsers,
