@@ -1,21 +1,19 @@
 import 'dart:io';
 
 import 'package:effektio/common/store/themes/SeperatedThemes.dart';
+import 'package:effektio/models/FaqModel.dart';
 import 'package:effektio/widgets/AppCommon.dart';
 import 'package:effektio/widgets/TagItem.dart';
-import 'package:effektio_flutter_sdk/effektio_flutter_sdk_ffi.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/painting.dart' as m_colors;
 import 'package:emoji_picker_flutter/emoji_picker_flutter.dart';
 
 class FaqItemScreen extends StatefulWidget {
-  final Client client;
-  final Faq faq;
+
+  final FaqModel faqModel;
 
   const FaqItemScreen({
-    Key? key,
-    required this.client,
-    required this.faq,
+    Key? key, required this.faqModel,
   }) : super(key: key);
 
   @override
@@ -36,7 +34,7 @@ class _FaqItemScreenState extends State<FaqItemScreen> {
 
     commentShowing = false;
     editFaqTitle = false;
-    faqController.text = widget.faq.title();
+    faqController.text = widget.faqModel.faq.title();
   }
 
   void onEmojiSelected(Emoji emoji) {
@@ -127,7 +125,7 @@ class _FaqItemScreenState extends State<FaqItemScreen> {
                         Flexible(
                           fit: FlexFit.loose,
                           child: Text(
-                            widget.faq.title(),
+                            widget.faqModel.faq.title(),
                             style: FAQTheme.titleStyle,
                           ),
                         ),
@@ -188,7 +186,7 @@ class _FaqItemScreenState extends State<FaqItemScreen> {
                               Padding(
                                 padding: const EdgeInsets.only(left: 8),
                                 child: Text(
-                                  widget.faq.likesCount().toString(),
+                                  widget.faqModel.faq.likesCount().toString(),
                                   style: FAQTheme.likeAndCommentStyle,
                                 ),
                               ),
@@ -209,7 +207,7 @@ class _FaqItemScreenState extends State<FaqItemScreen> {
                                     Padding(
                                       padding: const EdgeInsets.only(left: 8),
                                       child: Text(
-                                        widget.faq.commentsCount().toString(),
+                                        widget.faqModel.faq.commentsCount().toString(),
                                         style: FAQTheme.likeAndCommentStyle,
                                       ),
                                     ),
@@ -236,9 +234,9 @@ class _FaqItemScreenState extends State<FaqItemScreen> {
                     height: 40,
                     child: ListView.builder(
                       scrollDirection: Axis.horizontal,
-                      itemCount: widget.faq.tags().length,
+                      itemCount: widget.faqModel.faq.tags().length,
                       itemBuilder: (context, index) {
-                        var color = widget.faq.tags().elementAt(index).color();
+                        var color = widget.faqModel.faq.tags().elementAt(index).color();
                         var colorToShow = 0;
                         if (color != null) {
                           var colorList = color.rgbaU8();
@@ -251,7 +249,7 @@ class _FaqItemScreenState extends State<FaqItemScreen> {
                         }
 
                         return TagListItem(
-                          tagTitle: widget.faq.tags().elementAt(index).title(),
+                          tagTitle: widget.faqModel.faq.tags().elementAt(index).title(),
                           tagColor: colorToShow > 0
                               ? m_colors.Color(colorToShow)
                               : Colors.white,
