@@ -1,3 +1,4 @@
+import 'package:duration/duration.dart';
 import 'package:effektio/common/animations/LikeAnimation.dart';
 import 'package:effektio/common/store/themes/SeperatedThemes.dart';
 import 'package:effektio/widgets/NewsItem.dart';
@@ -31,7 +32,7 @@ class _NewsScreenState extends State<NewsScreen>
 
     controller = AnimationController(
       vsync: this,
-      duration: const Duration(seconds: 1),
+      duration: seconds(1),
     );
   }
 
@@ -55,7 +56,6 @@ class _NewsScreenState extends State<NewsScreen>
             ),
           );
         } else {
-          //final items = snapshot.requireData.toList();
           return Scaffold(
             extendBodyBehindAppBar: true,
             appBar: AppBar(
@@ -65,10 +65,15 @@ class _NewsScreenState extends State<NewsScreen>
                 builder: (BuildContext context) {
                   return IconButton(
                     icon: Container(
-                      margin: const EdgeInsets.only(bottom: 10, left: 10),
+                      margin: const EdgeInsets.only(
+                        bottom: 10,
+                        left: 10,
+                      ),
                       child: CircleAvatar(
                         backgroundColor: AppCommonTheme.primaryColor,
-                        child: Image.asset('assets/images/hamburger.png'),
+                        child: Image.asset(
+                          'assets/images/hamburger.png',
+                        ),
                       ),
                     ),
                     onPressed: () {
