@@ -1,28 +1,28 @@
-import 'dart:io';
-
+import 'package:beamer/beamer.dart';
 import 'package:effektio/common/store/themes/SeperatedThemes.dart';
 import 'package:effektio/controllers/chat_list_controller.dart';
-import 'package:effektio/screens/HomeScreens/chat/ChatScreen.dart';
+import 'package:effektio/controllers/receipt_controller.dart';
+import 'package:effektio/models/ChatModel.dart';
 import 'package:effektio/widgets/AppCommon.dart';
 import 'package:effektio/widgets/CustomAvatar.dart';
 import 'package:effektio_flutter_sdk/effektio_flutter_sdk_ffi.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:flutter_link_previewer/flutter_link_previewer.dart';
-import 'package:flutter_parsed_text/flutter_parsed_text.dart';
+import 'package:flutter_matrix_html/flutter_html.dart';
+import 'package:flutter_svg/svg.dart';
 import 'package:get/get.dart';
 import 'package:intl/intl.dart';
 
 class ChatListItem extends StatefulWidget {
-  final String userId;
+  final Client client;
   final Conversation room;
   final RoomMessage? latestMessage;
   final List<types.User> typingUsers;
 
   const ChatListItem({
     Key? key,
-    required this.userId,
+    required this.client,
     required this.room,
     this.latestMessage,
     required this.typingUsers,
@@ -35,6 +35,9 @@ class ChatListItem extends StatefulWidget {
 class _ChatListItemState extends State<ChatListItem> {
   Future<FfiBufferUint8>? avatar;
   String? displayName;
+  String? userId;
+
+  List<Member> activeMembers = [];
 
   @override
   void initState() {
@@ -50,6 +53,10 @@ class _ChatListItemState extends State<ChatListItem> {
         });
       }
     });
+
+    userId = widget.client.account().userId();
+
+    getActiveMembers();
   }
 
   @override
@@ -57,6 +64,7 @@ class _ChatListItemState extends State<ChatListItem> {
     String roomId = widget.room.getRoomId();
     // ToDo: UnreadCounter
     return Column(
+      mainAxisSize: MainAxisSize.min,
       children: <Widget>[
         ListTile(
           onTap: () => handleTap(context),
@@ -65,6 +73,8 @@ class _ChatListItemState extends State<ChatListItem> {
             avatar: avatar,
             displayName: displayName,
             radius: 25,
+            cacheHeight: 120,
+            cacheWidth: 120,
             isGroup: true,
             stringName: simplifyRoomId(roomId)!,
           ),
@@ -88,17 +98,12 @@ class _ChatListItemState extends State<ChatListItem> {
   }
 
   void handleTap(BuildContext context) {
-    Navigator.push(
-      context,
-      MaterialPageRoute(
-        builder: (context) => ChatScreen(
-          userId: widget.userId,
-          room: widget.room,
-          roomName: displayName,
-          roomAvatar: avatar,
-        ),
-      ),
-    );
+    Beamer.of(context).beamToNamed('/chat', data: ChatModel(
+      client: widget.client,
+      room: widget.room,
+      roomName: displayName,
+      roomAvatar: avatar,
+    ),);
   }
 
   Widget buildTitle(BuildContext context) {
@@ -129,72 +134,81 @@ class _ChatListItemState extends State<ChatListItem> {
     if (widget.latestMessage == null) {
       return const SizedBox();
     }
-    String sender = widget.latestMessage!.sender();
-    String body = widget.latestMessage!.body();
-    return Container(
-      margin: const EdgeInsets.symmetric(vertical: 10),
-      child: ParsedText(
-        text: '${simplifyUserId(sender)}: $body',
-        style: ChatTheme01.latestChatStyle,
-        regexOptions: const RegexOptions(multiLine: true, dotAll: true),
-        maxLines: 2,
-        parse: [
-          MatchText(
-            pattern: '(\\*\\*|\\*)(.*?)(\\*\\*|\\*)',
-            style: ChatTheme01.latestChatStyle.copyWith(
-              fontWeight: FontWeight.bold,
+    RoomEventItem? eventItem = widget.latestMessage!.eventItem();
+    if (eventItem == null) {
+      return const SizedBox();
+    }
+    String sender = eventItem.sender();
+    String itemContentType = eventItem.itemContentType();
+    if (itemContentType == 'Encrypted') {
+      return Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Padding(
+            padding: const EdgeInsets.symmetric(vertical: 10),
+            child: Text(
+              '${simplifyUserId(sender)}: ',
+              style: const TextStyle(color: ChatTheme01.chatBodyTextColor),
             ),
-            renderText: ({required String str, required String pattern}) {
-              return {'display': str.replaceAll(RegExp('(\\*\\*|\\*)'), '')};
-            },
-            onTap: (String value) => handleTap(context),
           ),
-          MatchText(
-            pattern: '_(.*?)_',
-            style: ChatTheme01.latestChatStyle.copyWith(
-              fontStyle: FontStyle.italic,
-            ),
-            renderText: ({required String str, required String pattern}) {
-              return {'display': str.replaceAll('_', '')};
-            },
-            onTap: (String value) => handleTap(context),
-          ),
-          MatchText(
-            pattern: '~(.*?)~',
-            style: ChatTheme01.latestChatStyle.copyWith(
-              decoration: TextDecoration.lineThrough,
-            ),
-            renderText: ({required String str, required String pattern}) {
-              return {'display': str.replaceAll('~', '')};
-            },
-            onTap: (String value) => handleTap(context),
-          ),
-          MatchText(
-            pattern: '`(.*?)`',
-            style: ChatTheme01.latestChatStyle.copyWith(
-              fontFamily: Platform.isIOS ? 'Courier' : 'monospace',
-            ),
-            renderText: ({required String str, required String pattern}) {
-              return {'display': str.replaceAll('`', '')};
-            },
-            onTap: (String value) => handleTap(context),
-          ),
-          MatchText(
-            pattern: regexEmail,
-            style: ChatTheme01.latestChatStyle.copyWith(
-              decoration: TextDecoration.underline,
-            ),
-            onTap: (String value) => handleTap(context),
-          ),
-          MatchText(
-            pattern: regexLink,
-            style: ChatTheme01.latestChatStyle.copyWith(
-              decoration: TextDecoration.underline,
-            ),
-            onTap: (String value) => handleTap(context),
+          const Text(
+            'RoomEncryptedEvent',
+            style: TextStyle(color: ChatTheme01.chatBodyTextColor),
           ),
         ],
-      ),
+      );
+    }
+    if (itemContentType == 'RedactedMessage') {
+      return Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Padding(
+            padding: const EdgeInsets.symmetric(vertical: 10),
+            child: Text(
+              '${simplifyUserId(sender)}: ',
+              style: const TextStyle(color: ChatTheme01.chatBodyTextColor),
+            ),
+          ),
+          const Text(
+            'RoomRedactionEvent',
+            style: TextStyle(color: ChatTheme01.chatBodyTextColor),
+          ),
+        ],
+      );
+    }
+    TextDesc? textDesc = eventItem.textDesc();
+    if (textDesc == null) {
+      return const SizedBox();
+    }
+    String body = textDesc.body();
+    String? formattedBody = textDesc.formattedBody();
+    if (formattedBody != null) {
+      body = simplifyBody(formattedBody);
+    }
+    return Row(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        Padding(
+          padding: const EdgeInsets.symmetric(vertical: 10),
+          child: Text(
+            '${simplifyUserId(sender)}: ',
+            style: const TextStyle(color: ChatTheme01.chatBodyTextColor),
+          ),
+        ),
+        Flexible(
+          child: Html(
+            padding: const EdgeInsets.symmetric(vertical: 10),
+            // ignore: unnecessary_string_interpolations
+            data: '''$body''',
+            maxLines: 1,
+            defaultTextStyle: const TextStyle(
+              color: ChatTheme01.chatBodyTextColor,
+              overflow: TextOverflow.ellipsis,
+            ),
+            onLinkTap: (url) => {},
+          ),
+        ),
+      ],
     );
   }
 
@@ -202,16 +216,65 @@ class _ChatListItemState extends State<ChatListItem> {
     if (widget.latestMessage == null) {
       return null;
     }
-    int? ts = widget.latestMessage!.originServerTs();
-    if (ts == null) {
+    RoomEventItem? eventItem = widget.latestMessage!.eventItem();
+    if (eventItem == null) {
       return null;
     }
-    return Text(
-      DateFormat.Hm().format(
-        DateTime.fromMillisecondsSinceEpoch(ts, isUtc: true),
-      ),
-      style: ChatTheme01.latestChatDateStyle,
+    String senderID = '';
+    types.Status? messageStatus;
+    int ts = eventItem.originServerTs();
+    var receiptController = Get.find<ReceiptController>();
+    List<String> seenByList = receiptController.getSeenByList(
+      widget.room.getRoomId(),
+      ts,
     );
+
+    senderID = widget.latestMessage!.eventItem()!.sender();
+
+    messageStatus = seenByList.isEmpty
+        ? types.Status.sent
+        : seenByList.length < activeMembers.length
+            ? types.Status.delivered
+            : types.Status.seen;
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: [
+        Text(
+          DateFormat.Hm().format(
+            DateTime.fromMillisecondsSinceEpoch(ts, isUtc: true),
+          ),
+          style: ChatTheme01.latestChatDateStyle,
+        ),
+        senderID == userId
+            ? customStatusBuilder(messageStatus)
+            : const SizedBox(),
+      ],
+    );
+  }
+
+  Widget customStatusBuilder(types.Status status) {
+    if (status == types.Status.delivered) {
+      return SvgPicture.asset('assets/images/deliveredIcon.svg');
+    } else if (status == types.Status.seen) {
+      return SvgPicture.asset('assets/images/seenIcon.svg');
+    } else if (status == types.Status.sending) {
+      return const Center(
+        child: SizedBox(
+          height: 10,
+          width: 10,
+          child: CircularProgressIndicator(
+            backgroundColor: Colors.transparent,
+            strokeWidth: 1.5,
+          ),
+        ),
+      );
+    } else {
+      return SvgPicture.asset(
+        'assets/images/sentIcon.svg',
+        width: 12,
+        height: 12,
+      );
+    }
   }
 
   String getUserPlural(List<types.User> authors) {
@@ -224,5 +287,9 @@ class _ChatListItemState extends State<ChatListItem> {
     } else {
       return '${authors[0].firstName} and ${authors.length - 1} others typing...';
     }
+  }
+
+  Future<void> getActiveMembers() async {
+    activeMembers = (await widget.room.activeMembers()).toList();
   }
 }
