@@ -1,22 +1,23 @@
+import 'package:beamer/beamer.dart';
 import 'package:effektio/common/store/themes/SeperatedThemes.dart';
 import 'package:effektio/controllers/chat_list_controller.dart';
-import 'package:effektio/screens/HomeScreens/chat/ChatScreen.dart';
+import 'package:effektio/models/ChatModel.dart';
 import 'package:effektio/widgets/AppCommon.dart';
 import 'package:effektio/widgets/CustomAvatar.dart';
 import 'package:effektio_flutter_sdk/effektio_flutter_sdk_ffi.dart'
-    show FfiBufferUint8, Invitation;
+    show Client, FfiBufferUint8, Invitation;
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:get/get.dart';
 
 class InviteInfoWidget extends StatefulWidget {
-  final String userId;
+  final Client client;
   final Invitation invitation;
   final Color avatarColor;
 
   const InviteInfoWidget({
     Key? key,
-    required this.userId,
+    required this.client,
     required this.invitation,
     required this.avatarColor,
   }) : super(key: key);
@@ -47,6 +48,7 @@ class _InviteInfoWidgetState extends State<InviteInfoWidget> {
 
   @override
   Widget build(BuildContext context) {
+    String userId = widget.client.userId().toString();
     return Card(
       color: AppCommonTheme.darkShade,
       margin: const EdgeInsets.symmetric(vertical: 1),
@@ -56,12 +58,12 @@ class _InviteInfoWidgetState extends State<InviteInfoWidget> {
           ListTile(
             // leading: CircleAvatar(backgroundColor: avatarColor),
             leading: CustomAvatar(
-              uniqueKey: widget.userId,
+              uniqueKey: userId,
               avatar: avatar,
               displayName: displayName,
               radius: 20,
               isGroup: true,
-              stringName: simplifyUserId(widget.userId)!,
+              stringName: simplifyUserId(userId)!,
             ),
             title: _buildTitle(),
             subtitle: _buildSubtitle(context),
@@ -122,15 +124,7 @@ class _InviteInfoWidgetState extends State<InviteInfoWidget> {
             final listController = Get.find<ChatListController>();
             for (var room in listController.joinedRooms) {
               if (room.conversation.getRoomId() == widget.invitation.roomId()) {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) => ChatScreen(
-                      userId: widget.userId,
-                      room: room.conversation,
-                    ),
-                  ),
-                );
+                Beamer.of(context).beamToNamed('/chat', data: ChatModel(client: widget.client, room: room.conversation));
               }
             }
           }
