@@ -1,10 +1,11 @@
+import 'package:beamer/beamer.dart';
 import 'package:effektio/common/store/themes/SeperatedThemes.dart';
 import 'package:effektio/widgets/AppCommon.dart';
 import 'package:effektio/widgets/CrossSigning.dart';
 import 'package:effektio/widgets/CustomAvatar.dart';
 import 'package:effektio_flutter_sdk/effektio_flutter_sdk.dart'
     show EffektioSdk;
-import 'package:effektio_flutter_sdk/effektio_flutter_sdk_ffi.dart' hide Color;
+import 'package:effektio_flutter_sdk/effektio_flutter_sdk_ffi.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_icons_null_safety/flutter_icons_null_safety.dart';
@@ -73,8 +74,8 @@ class SideDrawer extends StatelessWidget {
                 ),
               ),
               onPressed: () {
-                Navigator.pushNamed(context, '/login');
-              },
+                Beamer.of(context).beamToNamed('/login');
+                },
               child: Text(AppLocalizations.of(context)!.login),
             ),
           ),
@@ -88,14 +89,15 @@ class SideDrawer extends StatelessWidget {
                 ),
               ),
               onPressed: () {
-                Navigator.pushNamed(context, '/signup');
-              },
+                Beamer.of(context).beamToNamed('/signup');
+                },
               child: Text(AppLocalizations.of(context)!.signUp),
             ),
           ),
         ],
       );
     }
+
     return GestureDetector(
       onTap: () {
         showNotYetImplementedMsg(
@@ -110,6 +112,8 @@ class SideDrawer extends StatelessWidget {
             child: CustomAvatar(
               uniqueKey: userId,
               radius: 24,
+              cacheHeight: 120,
+              cacheWidth: 120,
               avatar: displayAvatar,
               displayName: displayName,
               isGroup: false,
@@ -163,7 +167,7 @@ class SideDrawer extends StatelessWidget {
         style: SideMenuAndProfileTheme.sideMenuStyle,
       ),
       onTap: () {
-        Navigator.pushNamed(context, '/todo');
+        Beamer.of(context).beamToNamed('/todo');
       },
     );
   }
@@ -181,7 +185,7 @@ class SideDrawer extends StatelessWidget {
         style: SideMenuAndProfileTheme.sideMenuStyle,
       ),
       onTap: () {
-        Navigator.pushNamed(context, '/gallery');
+        Beamer.of(context).beamToNamed('/gallery');
       },
     );
   }
@@ -349,7 +353,7 @@ class SideDrawer extends StatelessWidget {
         }
         final sdk = await EffektioSdk.instance;
         await sdk.logout();
-        Navigator.pushReplacementNamed(context, '/');
+        Beamer.of(context).beamToNamed('/');
       },
     );
   }
