@@ -1,7 +1,6 @@
-import 'package:flutter/foundation.dart';
-import 'package:flutter/services.dart';
-
 import 'package:effektio_flutter_sdk/effektio_flutter_sdk_platform_interface.dart';
+import 'package:flutter/services.dart';
+import 'package:meta/meta.dart';
 
 /// An implementation of [EffektioFlutterSdkPlatform] that uses method channels.
 class MethodChannelEffektioFlutterSdk extends EffektioFlutterSdkPlatform {
