@@ -1,23 +1,21 @@
 use anyhow::{bail, Context, Result};
 use clap::Parser;
-use effektio_core::{
-    events::{NewsContentType, NewsEventDevContent},
-    ruma::{
-        api::client::filter::RoomEventFilter,
-        events::{room::MediaSource, MessageLikeEvent},
-        RoomId,
-    },
-};
-use log::{info, warn};
 
 use crate::config::{LoginConfig, ENV_ROOM};
+use effektio_core::events;
+use effektio_core::ruma;
+
+use crate::ruma::api::client::filter::RoomEventFilter;
+use crate::ruma::events::room::MediaSource;
+
+use tracing::{info, warn};
 
 /// Posting a news item to a given room
 #[derive(Parser, Debug)]
 pub struct FetchNews {
     /// The room you want to post the news to
     #[clap(short, long, parse(try_from_str), env = ENV_ROOM)]
-    pub room: Box<RoomId>,
+    pub room: Box<ruma::RoomId>,
     #[clap(flatten)]
     pub login: LoginConfig,
 }
@@ -35,7 +33,7 @@ impl FetchNews {
         let mut query =
             matrix_sdk::room::MessagesOptions::backward().from(Some(sync_resp.next_batch.as_str()));
         let mut filter = RoomEventFilter::default();
-        filter.types = Some(types.as_slice());
+        filter.types = Some(types.to_vec());
         query.filter = filter;
         let messages = room.messages(query).await?;
         if messages.chunk.is_empty() {
@@ -44,10 +42,10 @@ impl FetchNews {
         for entry in messages.chunk {
             let event = match entry
                 .event
-                .deserialize_as::<MessageLikeEvent<NewsEventDevContent>>()
+                .deserialize_as::<ruma::events::MessageLikeEvent<events::NewsEventDevContent>>()
             {
-                Ok(MessageLikeEvent::Original(o)) => o,
-                Ok(MessageLikeEvent::Redacted(_)) => {
+                Ok(ruma::events::MessageLikeEvent::Original(o)) => o,
+                Ok(ruma::events::MessageLikeEvent::Redacted(_)) => {
                     // FIXME: what do we do with redactions
                     continue;
                 }
@@ -79,15 +77,15 @@ impl FetchNews {
             ]));
             for content in news.contents {
                 let (key, content) = match content {
-                    NewsContentType::Image(image) => match image.source {
+                    events::NewsContentType::Image(image) => match image.source {
                         MediaSource::Plain(url) => ("image", url.to_string()),
                         MediaSource::Encrypted(_) => ("image", "<encrypted>".to_owned()),
                     },
-                    NewsContentType::Video(video) => match video.source {
+                    events::NewsContentType::Video(video) => match video.source {
                         MediaSource::Plain(url) => ("video", url.to_string()),
                         MediaSource::Encrypted(_) => ("video", "<encrypted>".to_owned()),
                     },
-                    NewsContentType::Text(text) => ("text", text.body),
+                    events::NewsContentType::Text(text) => ("text", text.body),
                     _ => ("unknown", "n/a".to_owned()),
                 };
                 table.add_row(term_table::row::Row::new(vec![
