@@ -4,9 +4,9 @@ use clap::Parser;
 use crate::config::{LoginConfig, ENV_ROOM};
 use effektio_core::events;
 
-use log::info;
 use std::ffi::OsStr;
 use std::path::PathBuf;
+use tracing::info;
 
 /// Posting a news item to a given room
 #[derive(Parser, Debug)]
@@ -58,10 +58,10 @@ impl PostNews {
             };
             let image = std::fs::read(p).context("Couldn't open file for reading")?;
 
-            let response = client.media().upload(&mime, &image).await?;
+            let res = client.media().upload(&mime, image).await?;
 
             contents.push(events::NewsContentType::Image(
-                events::ImageMessageEventContent::plain("".to_owned(), response.content_uri, None),
+                events::ImageMessageEventContent::plain("".to_owned(), res.content_uri, None),
             ));
         }
 
