@@ -2,20 +2,18 @@ pub use matrix_sdk::ruma::{
     events::room::message::{
         ImageMessageEventContent, TextMessageEventContent, VideoMessageEventContent,
     },
+    events::RoomEventType,
     EventId,
 };
 
-mod comments;
+pub mod comments;
 mod common;
 mod labels;
 mod news;
-mod todos;
+pub mod tasks;
 
-pub use comments::{CommentEvent, CommentEventDevContent};
-pub use common::{BelongsTo, Color, Colorize, TimeZone, UtcDateTime};
+pub use common::{
+    BelongsTo, Color, Colorize, Reference, References, TimeZone, Update, UtcDateTime,
+};
 pub use labels::Labels;
 pub use news::{NewsContentType, NewsEvent, NewsEventDevContent};
-pub use todos::{
-    Priority as TaskPriority, SpecialTaskListRole, Task, TaskDevContent, TaskList,
-    TaskListDevContent,
-};
