@@ -1,14 +1,2 @@
-// #[cfg(test)]
-// mod auth;
-// #[cfg(test)]
-// mod formatted_body;
 #[cfg(test)]
-mod invitation;
-// #[cfg(test)]
-// mod news;
-// #[cfg(test)]
-// mod receipt;
-// #[cfg(test)]
-// mod typing;
-// #[cfg(test)]
-// mod verification;
+mod tests;
