@@ -7,10 +7,11 @@ async fn guest_can_login() -> Result<()> {
     let _ = env_logger::try_init();
     let tmp_dir = TempDir::new()?;
     let _client = guest_client(
-        tmp_dir.path().to_str().expect("always works").to_owned(),
+        tmp_dir.path().to_str().expect("always works").to_string(),
         option_env!("HOMESERVER")
             .unwrap_or("http://localhost:8118")
             .to_string(),
+        Some("GUEST_DEV".to_string()),
     )
     .await?;
     Ok(())
@@ -21,9 +22,10 @@ async fn sisko_can_login() -> Result<()> {
     let _ = env_logger::try_init();
     let tmp_dir = TempDir::new()?;
     let _client = login_new_client(
-        tmp_dir.path().to_str().expect("always works").to_owned(),
-        "@sisko:ds9.effektio.org".to_owned(),
-        "sisko".to_owned(),
+        tmp_dir.path().to_str().expect("always works").to_string(),
+        "@sisko:ds9.effektio.org".to_string(),
+        "sisko".to_string(),
+        Some("SISKO_DEV".to_string()),
     )
     .await?;
     Ok(())
@@ -34,9 +36,10 @@ async fn kyra_can_login() -> Result<()> {
     let _ = env_logger::try_init();
     let tmp_dir = TempDir::new()?;
     let _client = login_new_client(
-        tmp_dir.path().to_str().expect("always works").to_owned(),
-        "@kyra:ds9.effektio.org".to_owned(),
-        "kyra".to_owned(),
+        tmp_dir.path().to_str().expect("always works").to_string(),
+        "@kyra:ds9.effektio.org".to_string(),
+        "kyra".to_string(),
+        Some("KYRA_DEV".to_string()),
     )
     .await?;
     Ok(())
@@ -47,9 +50,10 @@ async fn kyra_can_restore() -> Result<()> {
     let _ = env_logger::try_init();
     let tmp_dir = TempDir::new()?;
     let client = login_new_client(
-        tmp_dir.path().to_str().expect("always works").to_owned(),
-        "@kyra:ds9.effektio.org".to_owned(),
-        "kyra".to_owned(),
+        tmp_dir.path().to_str().expect("always works").to_string(),
+        "@kyra:ds9.effektio.org".to_string(),
+        "kyra".to_string(),
+        Some("KYRA_DEV".to_string()),
     )
     .await?;
     let token = client.restore_token().await?;
@@ -57,7 +61,7 @@ async fn kyra_can_restore() -> Result<()> {
     drop(client);
 
     let client = login_with_token(
-        tmp_dir.path().to_str().expect("always works").to_owned(),
+        tmp_dir.path().to_str().expect("always works").to_string(),
         token,
     )
     .await?;
