@@ -8,59 +8,61 @@ async fn sisko_detects_kyra_read() -> Result<()> {
     let _ = env_logger::try_init();
 
     let tmp_dir = TempDir::new()?;
-    let sisko = login_new_client(
-        tmp_dir.path().to_str().expect("always works").to_owned(),
-        "@sisko:ds9.effektio.org".to_owned(),
-        "sisko".to_owned(),
+    let mut sisko = login_new_client(
+        tmp_dir.path().to_str().expect("always works").to_string(),
+        "@sisko:ds9.effektio.org".to_string(),
+        "sisko".to_string(),
+        Some("SISKO_DEV".to_string()),
     )
     .await?;
     let sisko_syncer = sisko.start_sync();
     let mut sisko_synced = sisko_syncer.first_synced_rx().expect("note yet read");
     while sisko_synced.next().await != Some(true) {} // let's wait for it to have synced
     let sisko_group = sisko
-        .get_group("#ops:ds9.effektio.org".to_owned())
+        .get_group("#ops:ds9.effektio.org".to_string())
         .await
         .expect("sisko should belong to ops");
     let event_id = sisko_group
-        .send_plain_message("Hi, everyone".to_owned())
+        .send_plain_message("Hi, everyone".to_string())
         .await?;
 
     let tmp_dir = TempDir::new()?;
-    let kyra = login_new_client(
-        tmp_dir.path().to_str().expect("always works").to_owned(),
-        "@kyra:ds9.effektio.org".to_owned(),
-        "kyra".to_owned(),
+    let mut kyra = login_new_client(
+        tmp_dir.path().to_str().expect("always works").to_string(),
+        "@kyra:ds9.effektio.org".to_string(),
+        "kyra".to_string(),
+        Some("KYRA_DEV".to_string()),
     )
     .await?;
     let kyra_syncer = kyra.start_sync();
     let mut first_synced = kyra_syncer.first_synced_rx().expect("note yet read");
     while first_synced.next().await != Some(true) {} // let's wait for it to have synced
     let kyra_group = kyra
-        .get_group("#ops:ds9.effektio.org".to_owned())
+        .get_group("#ops:ds9.effektio.org".to_string())
         .await
         .expect("kyra should belong to ops");
     kyra_group.read_receipt(event_id).await?;
 
-    let mut event_rx = kyra.receipt_event_rx()?;
+    let mut event_rx = kyra.receipt_event_rx().unwrap();
     loop {
         match event_rx.try_next() {
             Ok(Some(event)) => {
                 let mut found = false;
                 for record in event.receipt_records() {
-                    if record.user_id().as_str() == "@kyra:ds9.effektio.org" {
+                    if record.seen_by() == "@kyra:ds9.effektio.org" {
                         found = true;
                         break;
                     }
                 }
                 if found {
-                    println!("received: {:?}", event);
+                    println!("received: {event:?}");
                     break;
                 }
             }
             Ok(None) => {
                 println!("received: none");
             }
-            Err(e) => {}
+            Err(_e) => {}
         }
     }
 
