@@ -1,5 +1,6 @@
 use anyhow::Result;
 use futures::Stream;
+pub use futures_signals::signal_vec::VecDiff;
 use lazy_static::lazy_static;
 use tokio::runtime;
 
@@ -13,6 +14,8 @@ lazy_static! {
 mod account;
 mod auth;
 mod client;
+mod comments;
+mod common;
 mod conversation;
 mod device;
 mod group;
@@ -23,6 +26,7 @@ mod profile;
 mod receipt;
 mod room;
 mod stream;
+mod tasks;
 mod typing;
 mod verification;
 
@@ -30,17 +34,28 @@ pub use account::Account;
 pub use auth::{
     guest_client, login_new_client, login_with_token, register_with_registration_token,
 };
-pub use client::{Client, ClientStateBuilder, SyncState};
+pub use client::{Client, ClientStateBuilder, HistoryLoadState, SyncState};
+pub use comments::{Comment, CommentDraft, CommentsManager};
+pub use common::duration_from_secs;
 pub use conversation::{Conversation, CreateConversationSettingsBuilder};
+pub use core::time::Duration as EfkDuration;
 pub use device::{DeviceChangedEvent, DeviceLeftEvent, DeviceRecord};
-pub use effektio_core::models::{Color, Faq, News, Tag};
+pub use effektio_core::{
+    events::UtcDateTime,
+    models::{Color as EfkColor, Faq, News, Tag},
+};
 pub use group::{CreateGroupSettings, CreateGroupSettingsBuilder, Group};
 pub use invitation::Invitation;
-pub use message::{FileDescription, ImageDescription, RoomMessage};
+pub use message::{
+    FileDesc, ImageDesc, ReactionDesc, RoomEventItem, RoomMessage, RoomVirtualItem, TextDesc,
+};
 pub use profile::{RoomProfile, UserProfile};
 pub use receipt::{ReceiptEvent, ReceiptRecord};
 pub use room::{Member, Room};
-pub use stream::TimelineStream;
+pub use stream::{TimelineDiff, TimelineStream};
+pub use tasks::{
+    Task, TaskDraft, TaskList, TaskListDraft, TaskListUpdateBuilder, TaskUpdateBuilder,
+};
 pub use typing::TypingEvent;
 pub use verification::{VerificationEmoji, VerificationEvent};
 
