@@ -1,15 +1,8 @@
 use anyhow::{bail, Context, Result};
 use derive_builder::Builder;
 use effektio_core::statics::default_effektio_conversation_states;
-use futures::{
-    channel::mpsc::{channel, Receiver, Sender},
-    pin_mut, StreamExt,
-};
-use futures_signals::{
-    signal::{Mutable, MutableSignalCloned, SignalExt, SignalStream},
-    signal_vec::{SignalVecExt, VecDiff},
-};
-use js_int::uint;
+use futures::channel::mpsc::{channel, Receiver, Sender};
+use futures_signals::signal::{Mutable, MutableSignalCloned, SignalExt, SignalStream};
 use log::{error, info, warn};
 use matrix_sdk::{
     deserialized_responses::SyncTimelineEvent,
@@ -22,11 +15,13 @@ use matrix_sdk::{
         },
         assign,
         events::room::{
+            encrypted::OriginalSyncRoomEncryptedEvent,
             member::{MembershipState, OriginalSyncRoomMemberEvent},
             message::OriginalSyncRoomMessageEvent,
+            redaction::OriginalSyncRoomRedactionEvent,
         },
         serde::Raw,
-        OwnedRoomId, OwnedUserId, RoomId,
+        OwnedRoomId, OwnedUserId,
     },
     Client as MatrixClient,
 };
@@ -35,7 +30,7 @@ use std::sync::Arc;
 
 use super::{
     client::Client,
-    message::{sync_event_to_message, timeline_item_to_message, RoomMessage},
+    message::{sync_event_to_message, RoomMessage},
     receipt::ReceiptRecord,
     room::Room,
     RUNTIME,
@@ -73,7 +68,10 @@ impl Conversation {
         }
     }
 
-    fn set_latest_message(&mut self, msg: RoomMessage) {
+    fn set_latest_message(&mut self, mut msg: RoomMessage) {
+        if let Some(mut event_item) = msg.event_item() {
+            msg.set_event_item(Some(event_item));
+        }
         self.latest_message = Some(msg);
     }
 
@@ -110,13 +108,15 @@ impl std::ops::Deref for Conversation {
     }
 }
 
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 pub(crate) struct ConversationController {
     conversations: Mutable<Vec<Conversation>>,
     incoming_event_tx: Sender<RoomMessage>,
     incoming_event_rx: Arc<Mutex<Option<Receiver<RoomMessage>>>>,
+    encrypted_event_handle: Option<EventHandlerHandle>,
     message_event_handle: Option<EventHandlerHandle>,
     member_event_handle: Option<EventHandlerHandle>,
+    redacted_event_handle: Option<EventHandlerHandle>,
 }
 
 impl ConversationController {
@@ -126,14 +126,27 @@ impl ConversationController {
             conversations: Default::default(),
             incoming_event_tx: incoming_tx,
             incoming_event_rx: Arc::new(Mutex::new(Some(incoming_rx))),
+            encrypted_event_handle: None,
             message_event_handle: None,
             member_event_handle: None,
+            redacted_event_handle: None,
         }
     }
 
     pub fn add_event_handler(&mut self, client: &MatrixClient) {
         info!("sync room message event handler");
         let me = self.clone();
+
+        client.add_event_handler_context(me.clone());
+        let handle = client.add_event_handler(
+            |ev: Raw<OriginalSyncRoomEncryptedEvent>,
+             room: MatrixRoom,
+             c: MatrixClient,
+             Ctx(me): Ctx<ConversationController>| async move {
+                me.clone().process_room_encrypted(ev, &room, &c);
+            },
+        );
+        self.encrypted_event_handle = Some(handle);
 
         client.add_event_handler_context(me.clone());
         let handle = client.add_event_handler(
@@ -146,7 +159,7 @@ impl ConversationController {
         );
         self.message_event_handle = Some(handle);
 
-        client.add_event_handler_context(me);
+        client.add_event_handler_context(me.clone());
         let handle = client.add_event_handler(
             |ev: OriginalSyncRoomMemberEvent,
              room: MatrixRoom,
@@ -157,9 +170,24 @@ impl ConversationController {
             },
         );
         self.member_event_handle = Some(handle);
+
+        client.add_event_handler_context(me);
+        let handle = client.add_event_handler(
+            |ev: OriginalSyncRoomRedactionEvent,
+             room: MatrixRoom,
+             c: MatrixClient,
+             Ctx(me): Ctx<ConversationController>| async move {
+                me.clone().process_room_redaction(ev, &room, &c);
+            },
+        );
+        self.member_event_handle = Some(handle);
     }
 
     pub fn remove_event_handler(&mut self, client: &MatrixClient) {
+        if let Some(handle) = self.encrypted_event_handle.clone() {
+            client.remove_event_handler(handle);
+            self.encrypted_event_handle = None;
+        }
         if let Some(handle) = self.message_event_handle.clone() {
             client.remove_event_handler(handle);
             self.message_event_handle = None;
@@ -167,6 +195,10 @@ impl ConversationController {
         if let Some(handle) = self.member_event_handle.clone() {
             client.remove_event_handler(handle);
             self.member_event_handle = None;
+        }
+        if let Some(handle) = self.redacted_event_handle.clone() {
+            client.remove_event_handler(handle);
+            self.redacted_event_handle = None;
         }
     }
 
@@ -180,7 +212,52 @@ impl ConversationController {
         self.conversations.lock_mut().clone_from(&conversations);
     }
 
-    // this callback is called prior to load_rooms
+    // reorder room list on OriginalSyncRoomEncryptedEvent
+    async fn process_room_encrypted(
+        &mut self,
+        raw_event: Raw<OriginalSyncRoomEncryptedEvent>,
+        room: &MatrixRoom,
+        client: &MatrixClient,
+    ) {
+        info!("original sync room encrypted event: {:?}", raw_event);
+        if let MatrixRoom::Joined(joined) = room {
+            let mut convos = self.conversations.lock_mut();
+            let room_id = room.room_id();
+
+            let mut convo = Conversation::new(Room {
+                client: client.clone(),
+                room: room.clone(),
+            });
+            if let Ok(decrypted) = joined.decrypt_event(&raw_event).await {
+                let ev = raw_event
+                    .deserialize_as::<OriginalSyncRoomEncryptedEvent>()
+                    .unwrap();
+                let msg = RoomMessage::from_sync_event(
+                    None,
+                    None,
+                    ev.event_id.to_string(),
+                    ev.sender.to_string(),
+                    ev.origin_server_ts.get().into(),
+                    "Encrypted".to_string(),
+                    room,
+                    false,
+                );
+                convo.set_latest_message(msg.clone());
+
+                if let Some(idx) = convos.iter().position(|x| x.room_id() == room_id) {
+                    convos.remove(idx);
+                    convos.insert(0, convo);
+                    if let Err(e) = self.incoming_event_tx.try_send(msg) {
+                        warn!("Dropping ephemeral event for {}: {}", room_id, e);
+                    }
+                } else {
+                    convos.insert(0, convo);
+                }
+            }
+        }
+    }
+
+    // reorder room list on OriginalSyncRoomMessageEvent
     fn process_room_message(
         &mut self,
         ev: OriginalSyncRoomMessageEvent,
@@ -196,7 +273,16 @@ impl ConversationController {
                 client: client.clone(),
                 room: room.clone(),
             });
-            let msg = RoomMessage::from_original(&ev, room.clone());
+            let msg = RoomMessage::from_sync_event(
+                Some(ev.content.msgtype),
+                ev.content.relates_to,
+                ev.event_id.to_string(),
+                ev.sender.to_string(),
+                ev.origin_server_ts.get().into(),
+                "Message".to_string(),
+                room,
+                true,
+            );
             convo.set_latest_message(msg.clone());
 
             if let Some(idx) = convos.iter().position(|x| x.room_id() == room_id) {
@@ -211,6 +297,7 @@ impl ConversationController {
         }
     }
 
+    // reorder room list on OriginalSyncRoomMemberEvent
     fn process_room_member(
         &self,
         ev: OriginalSyncRoomMemberEvent,
@@ -252,6 +339,46 @@ impl ConversationController {
             }
         }
     }
+
+    // reorder room list on OriginalSyncRoomRedactionEvent
+    async fn process_room_redaction(
+        &mut self,
+        ev: OriginalSyncRoomRedactionEvent,
+        room: &MatrixRoom,
+        client: &MatrixClient,
+    ) {
+        info!("original sync room redaction event: {:?}", ev);
+        if let MatrixRoom::Joined(joined) = room {
+            let mut convos = self.conversations.lock_mut();
+            let room_id = room.room_id();
+
+            let mut convo = Conversation::new(Room {
+                client: client.clone(),
+                room: room.clone(),
+            });
+            let msg = RoomMessage::from_sync_event(
+                None,
+                None,
+                ev.event_id.to_string(),
+                ev.sender.to_string(),
+                ev.origin_server_ts.get().into(),
+                "RedactedMessage".to_string(),
+                room,
+                false,
+            );
+            convo.set_latest_message(msg.clone());
+
+            if let Some(idx) = convos.iter().position(|x| x.room_id() == room_id) {
+                convos.remove(idx);
+                convos.insert(0, convo);
+                if let Err(e) = self.incoming_event_tx.try_send(msg) {
+                    warn!("Dropping ephemeral event for {}: {}", room_id, e);
+                }
+            } else {
+                convos.insert(0, convo);
+            }
+        }
+    }
 }
 
 #[derive(Builder, Default, Clone)]
@@ -277,24 +404,24 @@ impl Client {
                 let initial_states = default_effektio_conversation_states();
                 let request = assign!(CreateRoomRequest::new(), {
                     creation_content: Some(Raw::new(&CreationContent::new())?),
-                    initial_state: &initial_states,
+                    initial_state: initial_states,
                     is_direct: true,
-                    invite: &settings.invites,
-                    room_alias_name: settings.alias.as_deref(),
-                    name: settings.name.as_ref().map(|x| x.as_ref()),
+                    invite: settings.invites,
+                    room_alias_name: settings.alias,
+                    name: settings.name,
                     visibility: Visibility::Private,
                 });
                 let response = client.create_room(request).await?;
-                Ok(response.room_id)
+                Ok(response.room_id().to_owned())
             })
             .await?
     }
 
-    pub(crate) async fn conversation(&self, name_or_id: String) -> Result<Conversation> {
+    pub async fn conversation(&self, name_or_id: String) -> Result<Conversation> {
         let me = self.clone();
         RUNTIME
             .spawn(async move {
-                if let Ok(room) = me.room(name_or_id) {
+                if let Ok(room) = me.room(name_or_id).await {
                     if !room.is_effektio_group().await {
                         Ok(Conversation::new(room))
                     } else {
