@@ -2,7 +2,7 @@ use anyhow::{bail, Context, Result};
 use futures_signals::signal::{
     Mutable, MutableSignal, MutableSignalCloned, SignalExt, SignalStream,
 };
-use log::{error, info, warn};
+use log::{error, info};
 use matrix_sdk::{
     event_handler::{Ctx, EventHandlerHandle},
     room::Room as MatrixRoom,
@@ -15,7 +15,11 @@ use matrix_sdk::{
 use std::time::{SystemTime, UNIX_EPOCH};
 use tokio::time::{sleep, Duration};
 
-use super::{client::Client, profile::UserProfile, RUNTIME};
+use super::{
+    client::{devide_groups_from_convos, Client},
+    profile::UserProfile,
+    RUNTIME,
+};
 
 #[derive(Default, Clone, Debug)]
 pub struct Invitation {
@@ -48,7 +52,7 @@ impl Invitation {
         let user_id = UserId::parse(self.sender.clone())?;
         RUNTIME
             .spawn(async move {
-                let mut user_profile = UserProfile::new(client, user_id);
+                let mut user_profile = UserProfile::new(client, user_id, None, None);
                 user_profile.fetch().await;
                 Ok(user_profile)
             })
@@ -126,7 +130,7 @@ impl Invitation {
     }
 }
 
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 pub(crate) struct InvitationController {
     invitations: Mutable<Vec<Invitation>>,
     stripped_event_handle: Option<EventHandlerHandle>,
@@ -207,7 +211,7 @@ impl InvitationController {
         client: &MatrixClient,
     ) -> Result<()> {
         // filter only event for me
-        let user_id = client.user_id().expect("You seem to be not logged in");
+        let user_id = client.user_id().context("You seem to be not logged in")?;
         if ev.state_key != *user_id {
             return Ok(());
         }
@@ -216,7 +220,7 @@ impl InvitationController {
         let start = SystemTime::now();
         let since_the_epoch = start
             .duration_since(UNIX_EPOCH)
-            .expect("Time went backwards");
+            .context("Time went backwards")?;
 
         info!("event type: StrippedRoomMemberEvent");
         info!("membership: {:?}", ev.content.membership);
@@ -285,5 +289,53 @@ impl Client {
             .invitations
             .signal_cloned()
             .to_stream()
+    }
+
+    pub async fn suggested_users_to_invite(&self, room_name: String) -> Result<Vec<UserProfile>> {
+        let client = self.clone();
+        let room_id = RoomId::parse(room_name)?;
+        let result = self.client.get_room(&room_id);
+        if result.is_none() {
+            return Ok(vec![]);
+        }
+        let room = result.unwrap();
+        RUNTIME
+            .spawn(async move {
+                // get member list of target room
+                let mut room_members = vec![];
+                let members = room.members().await?;
+                for member in members {
+                    room_members.push(member.user_id().to_string());
+                }
+                // iterate my rooms to get user list
+                let mut profiles: Vec<UserProfile> = vec![];
+                let (groups, convos) = devide_groups_from_convos(client.clone()).await;
+                for convo in convos {
+                    if convo.room_id() == room_id {
+                        continue;
+                    }
+                    let members = convo.members().await?;
+                    for member in members {
+                        let user_id = member.user_id().to_string();
+                        // exclude user that belongs to target room
+                        if room_members.contains(&user_id) {
+                            continue;
+                        }
+                        // exclude user that already selected
+                        if profiles.iter().any(|x| x.user_id() == member.user_id()) {
+                            continue;
+                        }
+                        let user_profile = UserProfile::new(
+                            client.client.clone(),
+                            member.user_id().to_owned(),
+                            member.avatar_url().map(|x| (*x).to_owned()),
+                            member.display_name().map(|x| x.to_string()),
+                        );
+                        profiles.push(user_profile);
+                    }
+                }
+                Ok(profiles)
+            })
+            .await?
     }
 }
