@@ -42,14 +42,14 @@ impl NewsDraft {
             .spawn(async move {
                 let mut image = std::fs::read(p).context("Couldn't open file for reading")?;
 
-                let response = me.client.media().upload(&mime, &image).await?;
+                let response = me.client.media().upload(&mime, image).await?;
 
                 let mut inner = me.content.lock_mut();
                 let counter = inner.contents.len();
 
                 inner.contents.push(events::NewsContentType::Image(
                     events::ImageMessageEventContent::plain(
-                        "".to_owned(),
+                        "".to_string(),
                         response.content_uri,
                         None,
                     ),
@@ -96,14 +96,14 @@ impl NewsDraft {
             .spawn(async move {
                 let mut image = std::fs::read(p).context("Couldn't open file for reading")?;
 
-                let response = me.client.media().upload(&mime, &image).await?;
+                let response = me.client.media().upload(&mime, image).await?;
 
                 let mut inner = me.content.lock_mut();
                 let counter = inner.contents.len();
 
                 inner.contents.push(events::NewsContentType::Image(
                     events::ImageMessageEventContent::plain(
-                        "".to_owned(),
+                        "".to_string(),
                         response.content_uri,
                         None,
                     ),
@@ -126,15 +126,15 @@ impl NewsDraft {
 }
 
 impl Group {
-    pub fn news_draft(&self) -> Result<NewsDraft> {
-        if let matrix_sdk::room::Room::Joined(joined) = &self.inner.room {
-            Ok(NewsDraft {
-                client: self.client.clone(),
-                room: joined.clone(),
-                content: Default::default(),
-            })
-        } else {
-            bail!("You can't create news for groups we are not part on")
-        }
-    }
+    // pub fn news_draft(&self) -> Result<NewsDraft> {
+    //     if let matrix_sdk::room::Room::Joined(joined) = &self.inner.room {
+    //         Ok(NewsDraft {
+    //             client: self.client.clone(),
+    //             room: joined.clone(),
+    //             content: Default::default(),
+    //         })
+    //     } else {
+    //         bail!("You can't create news for groups we are not part on")
+    //     }
+    // }
 }
