@@ -90,7 +90,7 @@ impl ReceiptEvent {
     }
 }
 
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 pub(crate) struct ReceiptController {
     event_tx: Sender<ReceiptEvent>,
     event_rx: Arc<Mutex<Option<Receiver<ReceiptEvent>>>>,
